requirements.txt
flask
mysql-connector-python
twilio
